net-print/hplip: Fix build with USE=-scanner
Closes: https://bugs.gentoo.org/973061

+++ net-print/hplip/hplip-3.25.8-r1.ebuild
@@ -263,7 +263,9 @@ src_install() {
 
 	# Installed by sane-backends
 	# Gentoo Bug: https://bugs.gentoo.org/show_bug.cgi?id=201023
-	rm "${ED}"/etc/sane.d/dll.conf || die
+	if use scanner ; then
+		rm "${ED}"/etc/sane.d/dll.conf || die
+	fi
 
 	rm "${ED}"/usr/share/doc/${PF}/{copyright,README_LIBJPG,COPYING} || die
 	rmdir --ignore-fail-on-non-empty "${ED}"/usr/share/doc/${PF}/ || die
